+++ src/views/index.ejs
